+++ src/components/Spotlight/Spotlight.tsx
@@ -54,6 +54,22 @@ export enum SpotlightShapes {
   'roundedBox' = 'rounded box',
 }
 
+// Recursively crawl the dom to find the nearest scrolling parent of the target element
+const findNearestScrollingParent = (el: HTMLElement | Element): HTMLElement | Element | null => {
+  const parent = el.parentElement;
+
+  if (parent && parent.tagName !== 'HTML') {
+    if (parent?.scrollHeight > parent?.clientHeight) {
+      // found it!
+      return parent;
+    }
+    return findNearestScrollingParent(parent);
+  }
+
+  // passing it back down
+  return null;
+};
+
 export type SpotlightProps = {
   StyledContainer?: StyledSubcomponentType;
   containerProps?: SubcomponentPropsType;
@@ -65,6 +81,7 @@ export type SpotlightProps = {
 
   children?: ReactNode;
   targetElement?: HTMLElement | Element;
+  scrollingParentElement?: HTMLElement | Element;
   backgroundBlur?: string;
   backgroundDarkness?: number;
   shape?: SpotlightShapes;
@@ -90,6 +107,7 @@ const Spotlight = ({
 
   children,
   targetElement,
+  scrollingParentElement, // this will get automatically picked if not defined
   backgroundBlur = '0.25rem',
   backgroundDarkness = 0.3,
   shape = SpotlightShapes.circular,
@@ -103,12 +121,18 @@ const Spotlight = ({
   scrollUpdateInterval = 0,
 }: SpotlightProps): JSX.Element | null => {
   const handleEventWithAnalytics = useAnalytics();
+  const scrollTarget = useRef(scrollingParentElement || null);
+
   const {
     width: windowWidth,
     height: windowHeight,
     isResizing,
-  } = useWindowSizeObserver(resizeUpdateInterval);
-  const { scrollY, isScrolling } = useScrollObserver(scrollUpdateInterval);
+  } = useWindowSizeObserver(resizeUpdateInterval, 50);
+
+  const { scrollY, isScrolling } = useScrollObserver(scrollUpdateInterval, 50, {
+    target: scrollTarget.current || undefined,
+  });
+
   const {
     performanceInfo: { tier: gpuTier },
     accessibilityPreferences: { prefersReducedMotion },
@@ -121,23 +145,48 @@ const Spotlight = ({
   const [isAutoScrolling, setIsAutoScrolling] = useState(false);
 
   useEffect(() => {
+    if (targetElement && !scrollTarget.current) {
+      scrollTarget.current = findNearestScrollingParent(targetElement);
+    }
+  }, [targetElement]);
+
+  useEffect(() => {
     if (targetElement) {
       const newTargetTop = targetElement?.getBoundingClientRect().top ?? 0;
       const annotationHeight = internalAnnotationRef?.current?.getBoundingClientRect().height;
-      // TODO: This assumes the annotation is above the target
+      // TODO: https://github.com/Headstorm/foundry-ui/issues/315
+      // This assumes the annotation is above the target
       // when custom above/under alignment is implemented, this logic should change
       // (if the annotation is below the target, this should be an addition, not subtraction)
       const offset = annotationHeight ? newTargetTop - annotationHeight : newTargetTop;
 
-      window.scrollBy({
-        top: offset,
-        left: 0,
-        behavior: !animateTargetChanges || prefersReducedMotion ? 'auto' : 'smooth',
-      });
+      if (scrollTarget.current) {
+        // TODO: https://github.com/Headstorm/foundry-ui/issues/483
+        // compare offset with the scrollTarget scrollHeight,
+        // if it's larger, subtract them and scroll the window next
+
+        scrollTarget.current.scrollTo({
+          top: offset,
+          left: 0,
+          behavior: !animateTargetChanges || prefersReducedMotion ? 'auto' : 'smooth',
+        });
+      } else {
+        window.scrollBy({
+          top: offset,
+          left: 0,
+          behavior: !animateTargetChanges || prefersReducedMotion ? 'auto' : 'smooth',
+        });
+      }
 
       setIsAutoScrolling(true);
     }
-  }, [animateTargetChanges, internalAnnotationRef, prefersReducedMotion, targetElement]);
+  }, [
+    animateTargetChanges,
+    internalAnnotationRef,
+    prefersReducedMotion,
+    scrollTarget,
+    targetElement,
+  ]);
 
   useEffect(() => {
     // check if auto scroll has completed
@@ -287,7 +336,7 @@ const Spotlight = ({
       lightPath: finalRectangularPath,
       circularLightPath: circularPath,
 
-      annotationTransform: `translate(${rect.x}px, ${rect.y}px) translate(0%, -100%)`, // TODO: change second translate to make different attach positions
+      annotationTransform: `translate(${rect.x}px, ${rect.y}px) translate(0%, -100%)`, // TODO: change second translate to make different attach positions https://github.com/Headstorm/foundry-ui/issues/315
 
       topBlurWidth: windowWidth,
       topBlurHeight: Math.max(rect.y - 1, 0),
@@ -338,7 +387,7 @@ const Spotlight = ({
     prefersReducedMotion,
   ]);
 
-  /* Click */
+  /* onClick */
 
   const handleClick = (evt: MouseEvent) => {
     // TODO: Rename to onClickOutside?